requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.2.4
djangorestframework==3.16.0
mysqlclient==2.2.7
pip==24.3.1
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions